agnes/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.39.0
pandas==2.2.0
numpy==1.26.4
statsmodels==0.14.1
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
